refactor(mempool-service): split service, coherence and subscriber into per-section modules
Move-only split of service.rs, coherence.rs and subscriber.rs along their
existing section dividers into per-section submodules, per the "sections are
modules" convention. No behavior, signature or public-API changes.
